MA-sim/requirements.txt
# Vietnamese Enterprise UBA Simulation Requirements
mysql-connector-python>=8.0.0
pandas>=1.3.0
numpy>=1.20.0
faker>=15.0.0
